Move getSortedOffers to utils
