+++ src/components/offers-list/offer-list.jsx
@@ -1,25 +1,10 @@
 import OfferCard from '../offer-card/offer-card';
 
-import OfferSortType from '../../consts/offer-sort-type';
+import getSortedOffers from '../../utils/get-sorted-offers';
 
 import offerProperties from "../../proptypes/offer-properties";
 
-const getSortedOffers = (offers, currentSortType) => {
-  switch (currentSortType) {
-    case OfferSortType.PRICE_LOW_TO_HIGH:
-      return offers.slice().sort((a, b) => a.price - b.price);
-    case OfferSortType.PRICE_HIGH_TO_LOW:
-      return offers.slice().sort((a, b) => b.price - a.price);
-    case OfferSortType.TOP_RATED:
-      return offers.slice().sort((a, b) => b.rating - a.rating);
-  }
-
-  return offers;
-};
-
-const OfferList = (props) => {
-
-  const {offers, handleCardHover, handleCardClick, sortingType, nearby} = props;
+const OfferList = ({offers, handleCardHover, handleCardClick, sortingType, nearby}) => {
 
   return (
     getSortedOffers(offers, sortingType).map((offer, index) =>
@@ -31,9 +16,7 @@ const OfferList = (props) => {
         nearby={nearby}
       />)
   );
-
 };
-
 
 OfferList.propTypes = {
   offers: PropTypes.arrayOf(PropTypes.shape(offerProperties)).isRequired,
